Exploit ParallelSearch into designAlongGrid.

+++ tests/test_siso_closed_loop_designer.py
@@ -1,23 +1,23 @@
 import controlSBML.siso_closed_loop_designer as cld
 from controlSBML.control_sbml import ControlSBML
 from controlSBML.siso_transfer_function_builder import SISOTransferFunctionBuilder
-import controlSBML as ctl
 import controlSBML.constants as cn
 from controlSBML.grid import Grid, Point
-from controlSBML.siso_design_evaluator import SISODesignEvaluator
 import helpers
 from controlSBML.timeseries import Timeseries
+from controlSBML.sbml_system import SBMLSystem
+from controlSBML.staircase import Staircase
+import controlSBML.util as util
 
 import copy
 import control
-import matplotlib.pyplot as plt
 import numpy as np
 import os
 import pandas as pd
 import sympy
 import unittest
 
-IGNORE_TEST = True
+IGNORE_TEST = False
 IS_PLOT = False
 FIGSIZE = (5, 5)
 helpers.setupPlotting(__file__)
@@ -68,7 +68,7 @@ end
 # Construct a transfer function for the model. This is a linear model, and so it should be accurate.
 INPUT_NAME = "S0"
 OUTPUT_NAME = "S2"
-SYSTEM = ctl.SBMLSystem(MODEL2, input_names=[INPUT_NAME], output_names=[OUTPUT_NAME], is_fixed_input_species=True)
+SYSTEM = SBMLSystem(MODEL2, input_names=[INPUT_NAME], output_names=[OUTPUT_NAME], is_fixed_input_species=True)
 TRANSFER_FUNCTION = control.TransferFunction(np.array([1.51083121, 2.01413339]), np.array([1.67214802, 1.24125478, 9.99999997]))
 TIMES = np.linspace(0, 20, 200)
 PARAMETER_DCT = {p: n+1 for n, p in enumerate(cn.CONTROL_PARAMETERS)}
@@ -101,6 +101,8 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
         self.remove()
 
     def remove(self):
+        if IGNORE_TEST:
+            return
         for ffile in REMOVE_FILES:
             if os.path.exists(ffile):
                 os.remove(ffile)
@@ -227,7 +229,7 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
         designer = self.makeDesigner()
         grid = Grid(min_value=0.1, max_value=10, num_coordinate=5)
         grid.addAxis("kp")
-        designer.designAlongGrid(grid, num_process=2)
+        designer.designAlongGrid(grid)
         self.assertGreater(designer.kp, 0)
         self.assertIsNone(designer.ki)
         self.assertIsNone(designer.kf)
@@ -237,9 +239,10 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
             return
         designer = self.makeDesigner()
         grid = Grid(min_value=0.1, max_value=10, num_coordinate=11)
+        grid = Grid(min_value=0.1, max_value=10, num_coordinate=5, is_random=False)
         grid.addAxis("kp")
         grid.addAxis("ki")
-        designer.designAlongGrid(grid, is_report=IGNORE_TEST, num_process=6)
+        designer.designAlongGrid(grid, is_report=IGNORE_TEST)
         self.assertGreater(designer.kp, 0)
         self.assertGreater(designer.ki, 0)
         self.assertIsNone(designer.kf)
@@ -247,18 +250,14 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
     def testPlotDesignResult(self):
         if IGNORE_TEST:
             return
-        self.init()
         designer = self.makeDesigner()
-        num_coordinate = 4
-        grid = Grid(min_value=0.1, max_value=2, num_coordinate=4, is_random=False)
-        grid.addAxis("kp", min_value=1, max_value=2, num_coordinate=num_coordinate)
-        grid.addAxis("ki", min_value=0.01, max_value=0.1, num_coordinate=num_coordinate)
-        grid.addAxis("kf", num_coordinate=num_coordinate)
-        designer.designAlongGrid(grid, num_process=6, is_report=IGNORE_TEST)
+        if not os.path.isfile(SAVE_PATH):
+            designer.design(kp_spec=True, ki_spec=True, kf_spec=True, min_value=0.1, max_value=10, 
+                 num_coordinate=5, num_restart=1, is_report=IGNORE_TEST)
         design_result_df = pd.read_csv(SAVE_PATH)
         def test(parameter_names):
             names = list(parameter_names)
-            names.append(cn.MSE)
+            names.append(cn.SCORE)
             designer._design_result_df = design_result_df[names]
             designer.plotDesignResult(is_plot=IS_PLOT, figsize=(5,5))
         #
@@ -286,17 +285,17 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
         func1 = lambda x: np.real(closed_loop_tf(x))
         cltf_nums = cltf.subs({kp: 2, ki: 3})
         func2 = lambda x: sympy.N(cltf_nums.subs({s: x}))
-        result = ctl.util.compareSingleArgumentFunctions(func1, func2, 0, 100)
+        result = util.compareSingleArgumentFunctions(func1, func2, 0, 100)
         self.assertTrue(result)
 
     def testBug1(self):
         if IGNORE_TEST:
             return
         # Setup
-        system = ctl.SBMLSystem(LINEAR_MDL, input_names=["k0"], output_names=["S3"], is_fixed_input_species=True,
+        system = SBMLSystem(LINEAR_MDL, input_names=["k0"], output_names=["S3"], is_fixed_input_species=True,
                                 is_steady_state=False)
         linear_bldr = SISOTransferFunctionBuilder(system)
-        linear_staircase = ctl.Staircase(initial_value=0, final_value=10, num_step=5)
+        linear_staircase = Staircase(initial_value=0, final_value=10, num_step=5)
         fitter_result = linear_bldr.fitTransferFunction(num_numerator=2, num_denominator=3, 
                                                     staircase=linear_staircase, fit_start_time=20,
                                                 start_time=0, end_time=200)
@@ -308,8 +307,8 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
         designer.evaluate(is_plot=IS_PLOT, figsize=FIGSIZE)
     
     def testBug3(self):
-        #if IGNORE_TEST:
-        #    return
+        if IGNORE_TEST:
+            return
         # Setup
         url = "https://www.ebi.ac.uk/biomodels/model/download/BIOMD0000000823.2?filename=Varusai2018.xml"
         INPUT_NAME = "pIRS"
@@ -317,10 +316,10 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
         ctlsb = ControlSBML(url, save_path=SAVE1_PATH)
         ctlsb.setOptions(input_names=[INPUT_NAME], output_names=[OUTPUT_NAME])
         #
-        grid = ctlsb.getGrid(kp_spec=True, ki_spec=False, num_coordinate=4, is_random=False)
+        grid = ctlsb.getGrid(kp_spec=True, ki_spec=False, num_coordinate=2, is_random=False)
         axis = grid.getAxis("kp")
         axis.setMinValue(0.1)
-        axis.setMaxValue(10)
+        axis.setMaxValue(1.1)
         ts, builder = ctlsb.plotGridDesign(grid, setpoint=120, num_restart=1, is_greedy=False, 
                                            is_plot=IS_PLOT, save_path=SAVE1_PATH)
         self.assertTrue(isinstance(ts, Timeseries))
@@ -340,90 +339,8 @@ class TestSISOClosedLoopDesigner(unittest.TestCase):
         axis.setMinValue(0.1)
         axis.setMaxValue(10)
         ts, builder = ctlsb.plotGridDesign(grid, setpoint=120, num_restart=1, is_plot=IS_PLOT,
-                                           num_process=-1, is_greedy=False, save_path=SAVE1_PATH)
+                                           is_greedy=False, save_path=SAVE1_PATH)
         self.assertTrue(isinstance(ts, Timeseries))
-
-    def testFindFeasibleClosedLoopSystem(self):
-        if IGNORE_TEST:
-            return
-        self.init()
-        times = np.linspace(0, 100, 1000)
-        designer = cld.SISOClosedLoopDesigner(self.system, self.sys_tf, times=times, setpoint=SETPOINT)
-        initial_dct = {"kp": 1, "ki": 1}
-        evaluator = SISODesignEvaluator(self.system, INPUT_NAME, OUTPUT_NAME, setpoint=SETPOINT, times=times)
-        value_dct = cld.SISOClosedLoopDesigner._searchForFeasibleClosedLoopSystem(evaluator, initial_dct)
-        designer.set(**value_dct)
-        ts, _ = designer.evaluate(is_plot=IS_PLOT)
-        self.assertTrue(np.isclose(SETPOINT, ts[OUTPUT_NAME].values[-1], atol=0.1))
-
-    def testEvaluatePoints(self):
-        if IGNORE_TEST:
-            return
-        points = np.repeat(Point(kp=1), 100)
-        self.init()
-        workunit = cld.Workunit(system=self.system.copy(), 
-                            input_name=INPUT_NAME,
-                            output_name=OUTPUT_NAME,
-                            setpoint=SETPOINT,
-                            times=TIMES,
-                            is_report=IGNORE_TEST,
-                            is_greedy=False,
-                            num_restart=1)
-        return_dct = {}
-        self.designer.evaluatePoints(0, 1, workunit, points, return_dct)
-        evaluator_result = return_dct[0]
-        self.assertEqual(len(evaluator_result[cn.MSE]), len(points))
-
-
-#############################
-class TestHistory(unittest.TestCase):
-
-    def setUp(self):
-        self.sys_tf = TRANSFER_FUNCTION
-        self.designer = cld.SISOClosedLoopDesigner(SYSTEM, self.sys_tf)
-        self.history = cld._History(self.designer)
-
-    def testConstructor(self):
-        if IGNORE_TEST:
-            return
-        self.assertEqual(len(self.history), 0)
-        diff = set(cn.CONTROL_PARAMETERS) - set(self.history._dct.keys())
-        self.assertEqual(len(diff), 0)
-
-    def addItems(self, count):
-        for _ in range(count):
-            self.history.add()
-    
-    def testAdd(self):
-        if IGNORE_TEST:
-            return
-        self.addItems(1)
-        self.assertEqual(len(self.history), 1)
-        self.addItems(1)
-        self.assertEqual(len(self.history), 2)
-
-    def testReport(self):
-        if IGNORE_TEST:
-            return
-        COUNT = 3
-        self.addItems(COUNT)
-        df = self.history.report()
-        self.assertEqual(len(df), COUNT)
-    
-    def testGet(self):
-        if IGNORE_TEST:
-            return
-        COUNT = 1
-        self.addItems(COUNT)
-        designer = self.history.get(0)
-        self.assertTrue(isinstance(designer, cld.SISOClosedLoopDesigner))
-
-    def testClear(self):
-        if IGNORE_TEST:
-            return
-        self.addItems(3)
-        self.history.clear()
-        self.assertEqual(len(self.history), 0)
 
 
 if __name__ == '__main__':
